requirements.txt
# Core dependencies
python-dotenv==1.0.0
pymongo==4.5.0
motor==3.3.2  # Async MongoDB driver

# Data extraction
aiohttp==3.9.1
asyncio==3.4.3
backoff==2.2.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
pandas==2.1.4
numpy==1.26.2
shapely==2.0.2  # For geometry processing
pyproj==3.6.1  # For coordinate transformations

# API and web scraping
requests==2.31.0
selenium==4.16.0  # If needed for dynamic content
webdriver-manager==4.0.1

# Utilities
python-dateutil==2.8.2
jsonschema==4.20.0
pydantic==2.5.2

# Logging and monitoring
colorlog==6.8.0
tqdm==4.66.1  # Progress bars

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1  # Code formatting
flake8==6.1.0  # Linting
ipython==8.18.1  # For debugging

# Optional: Machine Learning (for future enhancement)
# scikit-learn==1.3.2
# xgboost==2.0.3
# joblib==1.3.2
